+++ warp10/src/main/java/io/warp10/standalone/InMemoryChunkSet.java
@@ -222,7 +222,7 @@ public class InMemoryChunkSet {
    * @param timespan The timespan or value count to consider.
    * @return
    */
-  public GTSDecoder fetch(long now, long then, long count, long skip, double sample, CapacityExtractorOutputStream extractor, int preBoundary, int postBoundary) throws IOException {
+  public GTSDecoder fetch(long now, long then, long count, long skip, double sample, CapacityExtractorOutputStream extractor, long preBoundary, long postBoundary) throws IOException {
     GTSEncoder encoder = fetchEncoder(now, then, count, skip, sample, preBoundary, postBoundary);
 
     //
@@ -260,7 +260,7 @@ public class InMemoryChunkSet {
     return decoders;
   }
   
-  public GTSEncoder fetchEncoder(long now, long then, long count, long skip, double sample, int preBoundary, int postBoundary) throws IOException {
+  public GTSEncoder fetchEncoder(long now, long then, long count, long skip, double sample, long preBoundary, long postBoundary) throws IOException {
 
     if (this.ephemeral) {
       return fetchCountEncoder(Long.MAX_VALUE, 1L, postBoundary);
@@ -467,7 +467,7 @@ public class InMemoryChunkSet {
 //    return fetchCountEncoder(now, count).getUnsafeDecoder(false);
 //  }
 
-  private GTSEncoder fetchCountEncoder(long now, long count, int postBoundary) throws IOException {
+  private GTSEncoder fetchCountEncoder(long now, long count, long postBoundary) throws IOException {
     //
     // Determine the chunk id of 'now'
     // We offset it by chunkcount so we can safely decrement and
