+++ warp10/src/main/java/io/warp10/standalone/StandaloneMemoryStore.java
@@ -135,7 +135,7 @@ public class StandaloneMemoryStore extends Thread implements StoreClient {
   
   @Override
   //public GTSDecoderIterator fetch(final ReadToken token, final List<Metadata> metadatas, final long now, final long timespan, boolean fromArchive, boolean writeTimestamp, final int preBoundary, final int postBoundary) {
-  public GTSDecoderIterator fetch(final ReadToken token, final List<Metadata> metadatas, final long now, final long then, final long count, final long skip, final double sample, boolean writeTimestamp, final int preBoundary, final int postBoundary) {
+  public GTSDecoderIterator fetch(final ReadToken token, final List<Metadata> metadatas, final long now, final long then, final long count, final long skip, final double sample, boolean writeTimestamp, final long preBoundary, final long postBoundary) {
 
     if (0 != preBoundary || 0 != postBoundary) {
       throw new RuntimeException("Boundary retrieval is not supported by the current data store.");
